Create quest (#30)
* Moved dismissKeyboard to UIViewController extension, CreateQuestVC in progress, updated createQuest UI

* Added color picker, multiple changes.

* Updated prize core data model, added requestPassphraseAlertView UIViewController extension, fixed BananoQuest class, added quest data verification, fixed new wallet keyboard not dismissing.

* Added passphrase prompt at "Play now"

* added create quest map, added coreData constraint to questID, added context mergePolicy, questID is now based on quest index in the smart contract, added currentWallet static var to BananoQuest to maintain the wallet unlocked after using passphrase, other changes and fixes.

* Wallet flow from landing page reworked, fixed multiple issues, added handleWrongPassword, fixed create wallet flow, multiple changes.

+++ banano-quest/banano-quest/Models/Quest+CoreDataProperties.swift
@@ -25,7 +25,7 @@ extension Quest {
     @NSManaged public var merkleBody: String?
     @NSManaged public var maxWinners: Int16
     @NSManaged public var questID: Int32
-    @NSManaged public var prize: Int32
+    @NSManaged public var prize: Double
     @NSManaged public var winners: Winners?
     @NSManaged public var metadata: Metadata?
